requirements.txt
argcomplete==3.2.2
asgiref==3.7.2
boto3==1.34.49
botocore==1.34.49
certifi==2024.2.2
cfn-flip==1.3.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
durationpy==0.6
hjson==3.1.0
idna==2.10
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.5
mysqlclient==2.2.4
placebo==0.9.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.23.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.2
troposphere==4.6.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.25.11
Werkzeug==3.0.1
zappa==0.58.0
